fix(energy): address CodeRabbit review feedback
- Prevent lossy Number() conversion of on-chain space IDs in energy route
- Use direct eq() on normalized address columns (index-friendly) and trim+lowercase on write
- Validate device-id CSV in schema and parse strictly
- Render structured energy payloads instead of [object Object]
- Anchor energy proposal marker parsing to the trailing block (lastIndexOf)
- Reject array payloads before casting in proposal detail
- Gracefully disable energy actions when smart wallet is unavailable
- Translate energy agreement labels in de/es/fr/pt
- Widen factoryCommunityId to bigint in schema and migration
- Doc/NatSpec cleanups (EnergyToken, README fences, interval data feed)

+++ packages/storage-postgres/migrations/0053_energy_communities.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS "energy_communities" (
 	"community_proxy_address" text NOT NULL,
 	"energy_token_address" text NOT NULL,
 	"admin_address" text NOT NULL,
-	"factory_community_id" integer,
+	"factory_community_id" bigint,
 	"activated_at" timestamp NOT NULL,
 	"created_at" timestamp DEFAULT now() NOT NULL,
 	"updated_at" timestamp DEFAULT now() NOT NULL
